safari-600.7-branch/LayoutTests/svg/custom/svg-parse-overflow-3-expected.txt

CONSOLE MESSAGE: line 14: Error: Invalid value for <rect> attribute height="105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
